Replacing wrong balun with correct part
We used some 1:4 part, but we really should use a 1:1 part with
center tap. Otherwise the impedance seen when looking into the
atenna port will be way off. What we really need to see is the
sampling cap (a short to GND at RF frequencies), followed by the
internal switch resistance (5 ish Ohm or so?), followed by the
10 Ohm series resistor, i.e. a total of 15ish Ohm multiplied by
4 (60 Ohm). If we use a 1:1 transformer, then one secondary winding from
the center tap to the end has half the turns of the primary side;
a turns ratio of 2:1 yields an impedance ratio of 4:1 which is
what we are looking for.